requirements.txt
Brotli==1.1.0
inflate64==1.0.3
multivolumefile==0.2.3
numpy==2.2.3
pandas==2.2.3
psutil==7.0.0
py7zr==1.0.0
pybcj==1.0.6
pycryptodomex==3.23.0
pyppmd==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.1
pyzstd==0.17.0
six==1.17.0
texttable==1.7.0
tzdata==2025.1
Pillow==10.4.0
sqlalchemy